Refactored LoginForm to one file and deleted extra files
